requirements.txt
langchain-community
langchain-core
langchain-openai
python-dotenv
langchain
jq
